backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
PyYAML
passlib[bcrypt]
PyJWT>=2.11,<3
python-multipart
bcrypt==4.0.1
alembic>=1.13,<2
slowapi>=0.1.9,<1
